Boot from another OSes

+++ src/os/include/uefi/ueficonsolecontrolscreenmode.h
@@ -7,7 +7,7 @@
 
 
 
-enum class UefiConsoleControlScreenMode: u8
+enum class UefiConsoleControlScreenMode: u32
 {
     TEXT,
     GRAPHICS,
